Add locale success message on the Wufoo form
We are showing only a portuguese message.

This patch sends a hidden sucess message on each form, this way the user will
receive the message on your locale.

This field is hidden in the Wufoo and has the default value the portuguese
success mesage, so if the user access the form via Wufoo interface the
experience will be the same.